Switch token_infos_ to a ValueStore
